refactor: changed some folder paths
