+++ src/layout/admin-layout.tsx
@@ -15,7 +15,7 @@ import Shadow from "@/components/ui/shadow";
 import { useLayoutEffect, useMemo } from "react";
 import { cn } from "@/lib/utils";
 import AppRoutes from "@/constants/app-routes";
-import Sidebar from "../common/sidebar/sideBar";
+import Sidebar from "../components/common/sidebar/sideBar";
 
 // Interface for Props
 interface PropsType {
