+++ src/routes/instructor-routes.tsx
@@ -1,4 +1,4 @@
-import InstructorLayout from "@/components/layout/instructor-layout";
+import InstructorLayout from "@/layout/instructor-layout";
 import Dashboard from "@/pages/instructor/dashboard";
 import Login from "@/pages/authentication/login";
 import { useState } from "react";
@@ -7,7 +7,7 @@ import PublicRoutes from "@/routes/public-routes";
 import ProtectedRoutes from "@/routes/protected-routes";
 import ForgotPassword from "@/pages/authentication/forgot-password";
 import Reset from "@/pages/authentication/reset-password";
-import AuthenticationLayout from "@/components/layout/authentication-layout";
+import AuthenticationLayout from "@/layout/authentication-layout";
 import AppRoutes from "@/constants/app-routes";
 import Chat from "@/pages/chat/chat";
 import Profile from "@/pages/profile/profile";
